backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apturl==0.5.2
bcrypt==4.0.1
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==8.10.0
motor==3.7.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
orjson==3.11.4
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.5.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.6.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.15.3
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
SecretStorage==3.3.1
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.49.0
systemd-python==234
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.11.0
unattended-upgrades==0.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.3
wadllib==1.3.6
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
